Make common Uri class for the deeplinks support
