+++ navigation/navigation-common/src/commonTest/kotlin/androidx/navigation/NavDeepLinkActionTest.kt
@@ -16,13 +16,10 @@
 
 package androidx.navigation
 
-import android.net.Uri
-import androidx.test.filters.SmallTest
-import com.google.common.truth.Truth.assertThat
-import com.google.common.truth.Truth.assertWithMessage
-import org.junit.Test
+import androidx.kruth.assertThat
+import androidx.kruth.assertWithMessage
+import kotlin.test.Test
 
-@SmallTest
 class NavDeepLinkActionTest {
 
     companion object {
@@ -36,7 +33,7 @@ class NavDeepLinkActionTest {
 
         assertWithMessage("The actions should not have matched")
             .that(
-                deepLink.matches(NavDeepLinkRequest(Uri.parse(DEEP_LINK_EXACT_HTTPS), null, null))
+                deepLink.matches(NavDeepLinkRequest(UriUtils.parse(DEEP_LINK_EXACT_HTTPS), null, null))
             )
             .isFalse()
     }
@@ -48,7 +45,7 @@ class NavDeepLinkActionTest {
         assertWithMessage("The actions should have matched")
             .that(
                 deepLink.matches(
-                    NavDeepLinkRequest(Uri.parse(DEEP_LINK_EXACT_HTTPS), DEEP_LINK_ACTION, null)
+                    NavDeepLinkRequest(UriUtils.parse(DEEP_LINK_EXACT_HTTPS), DEEP_LINK_ACTION, null)
                 )
             )
             .isTrue()
