+++ navigation/navigation-common/src/commonTest/kotlin/androidx/navigation/NavDeepLinkMimeTypeTest.kt
@@ -16,14 +16,11 @@
 
 package androidx.navigation
 
-import android.net.Uri
-import androidx.test.filters.SmallTest
-import com.google.common.truth.Truth.assertThat
-import com.google.common.truth.Truth.assertWithMessage
-import org.junit.Assert.fail
-import org.junit.Test
+import androidx.kruth.assertThat
+import androidx.kruth.assertWithMessage
+import kotlin.test.Test
+import kotlin.test.fail
 
-@SmallTest
 class NavDeepLinkMimeTypeTest {
     companion object {
         private const val DEEP_LINK_EXACT_HTTPS = "https://www.example.com"
@@ -63,7 +60,7 @@ class NavDeepLinkMimeTypeTest {
         assertWithMessage("The mimeTypes should match")
             .that(
                 deepLink.matches(
-                    NavDeepLinkRequest(Uri.parse(DEEP_LINK_EXACT_HTTPS), null, mimeType)
+                    NavDeepLinkRequest(UriUtils.parse(DEEP_LINK_EXACT_HTTPS), null, mimeType)
                 )
             )
             .isTrue()
