+++ navigation/navigation-common/src/commonTest/kotlin/androidx/navigation/NavDeepLinkTest.kt
@@ -34,11 +34,12 @@ import androidx.navigation.test.stringArrayArgument
 import androidx.navigation.test.stringListArgument
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 
-// TODO: Move to commonTest once Uri overloads will be commonized
 class NavDeepLinkTest {
+
+    @Serializable class TestClass
 
     companion object {
         private const val DEEP_LINK_EXACT_NO_SCHEME = "www.example.com"
@@ -51,10 +52,10 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(null, "test.action", null)
 
         assertWithMessage("NavDeepLink shouldn't match with null Uri")
-            .that(deepLink.matches(DEEP_LINK_EXACT_HTTP))
+            .that(deepLink.matches(UriUtils.parse(DEEP_LINK_EXACT_HTTP)))
             .isFalse()
         assertWithMessage("NavDeepLink shouldn't find matching arguments with null Uri")
-            .that(deepLink.getMatchingArguments(DEEP_LINK_EXACT_HTTP, mapOf()))
+            .that(deepLink.getMatchingArguments(UriUtils.parse(DEEP_LINK_EXACT_HTTP), mapOf()))
             .isNull()
     }
 
@@ -63,10 +64,10 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(DEEP_LINK_EXACT_HTTP)
 
         assertWithMessage("HTTP link should match HTTP")
-            .that(deepLink.matches(DEEP_LINK_EXACT_HTTP))
+            .that(deepLink.matches(UriUtils.parse(DEEP_LINK_EXACT_HTTP)))
             .isTrue()
         assertWithMessage("HTTP link should not match HTTPS")
-            .that(deepLink.matches(DEEP_LINK_EXACT_HTTPS))
+            .that(deepLink.matches(UriUtils.parse(DEEP_LINK_EXACT_HTTPS)))
             .isFalse()
     }
 
@@ -75,8 +76,7 @@ class NavDeepLinkTest {
         val deepLinkString = "android-app://com.example"
         val deepLink = NavDeepLink(deepLinkString)
 
-        assertThat(deepLink.matches(deepLinkString))
-            .isTrue()
+        assertThat(deepLink.matches(UriUtils.parse(deepLinkString))).isTrue()
     }
 
     @Test
@@ -84,8 +84,7 @@ class NavDeepLinkTest {
         val deepLinkString = "android+app://com.example"
         val deepLink = NavDeepLink(deepLinkString)
 
-        assertThat(deepLink.matches(deepLinkString))
-            .isTrue()
+        assertThat(deepLink.matches(UriUtils.parse(deepLinkString))).isTrue()
     }
 
     @Test
@@ -93,8 +92,7 @@ class NavDeepLinkTest {
         val deepLinkString = "android.app://com.example"
         val deepLink = NavDeepLink(deepLinkString)
 
-        assertThat(deepLink.matches(deepLinkString))
-            .isTrue()
+        assertThat(deepLink.matches(UriUtils.parse(deepLinkString))).isTrue()
     }
 
     @Test
@@ -102,10 +100,10 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(DEEP_LINK_EXACT_NO_SCHEME)
 
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(DEEP_LINK_EXACT_HTTP))
+            .that(deepLink.matches(UriUtils.parse(DEEP_LINK_EXACT_HTTP)))
             .isTrue()
         assertWithMessage("No scheme deep links should match https")
-            .that(deepLink.matches(DEEP_LINK_EXACT_HTTPS))
+            .that(deepLink.matches(UriUtils.parse(DEEP_LINK_EXACT_HTTPS)))
             .isTrue()
     }
 
@@ -115,13 +113,9 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = "2"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id),
-            mapOf()
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(UriUtils.parse(deepLinkArgument.replace("{id}", id)), mapOf())
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the id")
             .that(matchArgs?.getString("id"))
             .isEqualTo(id)
@@ -133,16 +127,13 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{id}", id.toString())),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -151,13 +142,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = "invalid"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should be null")
-            .that(matchArgs)
-            .isNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{id}", id)),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should be null").that(matchArgs).isNull()
     }
 
     @Test
@@ -167,19 +157,15 @@ class NavDeepLinkTest {
 
         val id = 2
         val myArg = "test"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id.toString()).replace("{myarg}", myArg),
-            mapOf(
-                "id" to intArgument(),
-                "myarg" to stringArgument()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    deepLinkArgument.replace("{id}", id.toString()).replace("{myarg}", myArg)
+                ),
+                mapOf("id" to intArgument(), "myarg" to stringArgument())
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
         assertWithMessage("Args should contain the argument")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo(myArg)
@@ -193,19 +179,15 @@ class NavDeepLinkTest {
 
         val id = 211
         val myArg = "test"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id.toString()).replace("{myarg}", myArg),
-            mapOf(
-                "id" to intArgument(),
-                "myarg" to stringArgument()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    deepLinkArgument.replace("{id}", id.toString()).replace("{myarg}", myArg)
+                ),
+                mapOf("id" to intArgument(), "myarg" to stringArgument())
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
         assertWithMessage("Args should contain the argument")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo(myArg)
@@ -219,16 +201,13 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 211
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{id}", id.toString())),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -238,7 +217,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument.replace("{name}", ""),
+                UriUtils.parse(deepLinkArgument.replace("{name}", "")),
                 mapOf("name" to stringArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -254,7 +233,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument.replace("{name}", ""),
+                UriUtils.parse(deepLinkArgument.replace("{name}", "")),
                 mapOf("name" to intArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNull()
@@ -267,7 +246,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument.replace("{name}", ""),
+                UriUtils.parse(deepLinkArgument.replace("{name}", "")),
                 mapOf("name" to stringArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -283,7 +262,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument.replace("{name}", ""),
+                UriUtils.parse(deepLinkArgument.replace("{name}", "")),
                 mapOf("name" to stringArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -299,7 +278,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument.replace("{first}", "").replace("{last}", "last"),
+                UriUtils.parse(deepLinkArgument.replace("{first}", "").replace("{last}", "last")),
                 mapOf("first" to stringArgument(), "last" to stringArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -318,10 +297,12 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument
-                    .replace("{first}", "first")
-                    .replace("{centre}", "")
-                    .replace("{last}", "last"),
+                UriUtils.parse(
+                    deepLinkArgument
+                        .replace("{first}", "first")
+                        .replace("{centre}", "")
+                        .replace("{last}", "last")
+                ),
                 mapOf(
                     "first" to stringArgument(),
                     "centre" to stringArgument(),
@@ -347,7 +328,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument.replace("{first}", "first").replace("{last}", ""),
+                UriUtils.parse(deepLinkArgument.replace("{first}", "first").replace("{last}", "")),
                 mapOf("first" to stringArgument(), "last" to stringArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -366,10 +347,12 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument
-                    .replace("{first}", "")
-                    .replace("{centre}", "centre")
-                    .replace("{last}", ""),
+                UriUtils.parse(
+                    deepLinkArgument
+                        .replace("{first}", "")
+                        .replace("{centre}", "centre")
+                        .replace("{last}", "")
+                ),
                 mapOf(
                     "first" to stringArgument(),
                     "centre" to stringArgument(),
@@ -395,10 +378,12 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument
-                    .replace("{first}", "")
-                    .replace("{centre}", "")
-                    .replace("{last}", ""),
+                UriUtils.parse(
+                    deepLinkArgument
+                        .replace("{first}", "")
+                        .replace("{centre}", "")
+                        .replace("{last}", "")
+                ),
                 mapOf(
                     "first" to stringArgument(),
                     "centre" to stringArgument(),
@@ -424,10 +409,12 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument
-                    .replace("{first}", "")
-                    .replace("{centre}", "")
-                    .replace("{last}", ""),
+                UriUtils.parse(
+                    deepLinkArgument
+                        .replace("{first}", "")
+                        .replace("{centre}", "")
+                        .replace("{last}", "")
+                ),
                 mapOf(
                     "first" to stringArgument(),
                     "centre" to stringArgument(),
@@ -453,10 +440,12 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument
-                    .replace("{first}", "")
-                    .replace("{centre}", "")
-                    .replace("{last}", ""),
+                UriUtils.parse(
+                    deepLinkArgument
+                        .replace("{first}", "")
+                        .replace("{centre}", "")
+                        .replace("{last}", "")
+                ),
                 mapOf(
                     "first" to stringArgument(),
                     "centre" to stringArgument(),
@@ -482,7 +471,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument.replace("{first}", "").replace("{last}", "last"),
+                UriUtils.parse(deepLinkArgument.replace("{first}", "").replace("{last}", "last")),
                 mapOf("first" to stringArgument(), "last" to stringArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -501,7 +490,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument.replace("{first}", "").replace("{last}", ""),
+                UriUtils.parse(deepLinkArgument.replace("{first}", "").replace("{last}", "")),
                 mapOf("first" to stringArgument(), "last" to stringArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -521,16 +510,15 @@ class NavDeepLinkTest {
 
         val id = 2
         val myArg = "test"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id.toString()).replace("{myarg}", myArg),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    deepLinkArgument.replace("{id}", id.toString()).replace("{myarg}", myArg)
+                ),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
         assertWithMessage("Args should contain the argument")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo(myArg)
@@ -542,16 +530,13 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{id}", id.toString())),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -560,13 +545,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = "invalid"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should be null")
-            .that(matchArgs)
-            .isNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{id}", id)),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should be null").that(matchArgs).isNull()
     }
 
     @Test
@@ -576,20 +560,15 @@ class NavDeepLinkTest {
 
         val id = 2
         val myarg = "test"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument
-                .replace("{id}", id.toString()).replace("{myarg}", myarg),
-            mapOf(
-                "id" to intArgument(),
-                "myarg" to stringArgument()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    deepLinkArgument.replace("{id}", id.toString()).replace("{myarg}", myarg)
+                ),
+                mapOf("id" to intArgument(), "myarg" to stringArgument())
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
         assertWithMessage("Args should contain the argument")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo(myarg)
@@ -601,13 +580,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("id" to intArgument(id))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("id" to intArgument(id))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the id")
             .that(matchArgs?.containsKey("id"))
             .isFalse()
@@ -618,13 +596,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("myarg" to nullableStringArgument(null))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("myarg" to nullableStringArgument(null))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the argument")
             .that(matchArgs?.containsKey("myarg"))
             .isFalse()
@@ -637,13 +614,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument,
-            mapOf("id" to intArgument(id))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument),
+                mapOf("id" to intArgument(id))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the id")
             .that(matchArgs?.containsKey("id"))
             .isFalse()
@@ -655,16 +631,13 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val arg = ""
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{arg}", arg),
-            mapOf("arg" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.containsKey("arg"))
-            .isTrue()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{arg}", arg)),
+                mapOf("arg" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.containsKey("arg")).isTrue()
     }
 
     // Ensure case when matching the exact argument query (i.e. param names in braces) is handled
@@ -673,13 +646,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument,
-            mapOf("myarg" to nullableStringArgument(null))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument),
+                mapOf("myarg" to nullableStringArgument(null))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         // We allow {argName} values for String types
         assertWithMessage("Args should contain the argument")
             .that(matchArgs?.getString("myarg"))
@@ -694,7 +666,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument,
+                UriUtils.parse(deepLinkArgument),
                 mapOf("myarg" to stringListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -710,18 +682,19 @@ class NavDeepLinkTest {
     fun deepLinkQueryParamNullableNonStringArgumentMatchParamsInBraces() {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
-        val intArrayArg = NavArgument.Builder().setType(NavType.IntArrayType)
-            .setIsNullable(true)
-            .setDefaultValue(null)
-            .build()
+        val intArrayArg =
+            NavArgument.Builder()
+                .setType(NavType.IntArrayType)
+                .setIsNullable(true)
+                .setDefaultValue(null)
+                .build()
 
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument,
-            mapOf("myarg" to intArrayArg)
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument),
+                mapOf("myarg" to intArrayArg)
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         // For non-strings, {argName} values are invalid and considered lack of argument value
         assertWithMessage("Args should not contain the argument")
             .that(matchArgs?.containsKey("myarg"))
@@ -738,7 +711,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument,
+                UriUtils.parse(deepLinkArgument),
                 mapOf("myarg" to intArrayArg)
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -763,7 +736,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                deepLinkArgument,
+                UriUtils.parse(deepLinkArgument),
                 mapOf("myarg" to intArrayArg)
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -775,16 +748,18 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{myarg}", "myarg"),
-            mapOf("myarg" to NavArgument.Builder()
-                .setType(NavType.StringType)
-                .setIsNullable(true)
-                .build())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{myarg}", "myarg")),
+                mapOf(
+                    "myarg" to
+                        NavArgument.Builder()
+                            .setType(NavType.StringType)
+                            .setIsNullable(true)
+                            .build()
+                )
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the argument and it should not be null")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo("myarg")
@@ -797,19 +772,13 @@ class NavDeepLinkTest {
 
         val id = 2
         val optional = "test"
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?id={id}".replace("{id}", id.toString()),
-            mapOf(
-                "id" to intArgument(),
-                "optional" to stringArgument(optional)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?id={id}".replace("{id}", id.toString())),
+                mapOf("id" to intArgument(), "optional" to stringArgument(optional))
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
         assertWithMessage("Args should not contain optional")
             .that(matchArgs?.containsKey("optional"))
             .isFalse()
@@ -821,20 +790,16 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?optional={optional}&id={id}"
-                .replace("{id}", id.toString()),
-            mapOf(
-                "id" to intArgument(),
-                "optional" to stringArrayArgument(arrayOf("theArg"))
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    "$DEEP_LINK_EXACT_HTTPS/users?optional={optional}&id={id}"
+                        .replace("{id}", id.toString())
+                ),
+                mapOf("id" to intArgument(), "optional" to stringArrayArgument(arrayOf("theArg")))
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
         assertWithMessage("Args should not contain optional")
             .that(matchArgs?.getStringArray("optional"))
             .isEqualTo(arrayOf("{optional}"))
@@ -846,19 +811,13 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?id={id}".replace("{id}", id.toString()),
-            mapOf(
-                "id" to intArgument(),
-                "optional" to nullableStringArgument(null)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?id={id}".replace("{id}", id.toString())),
+                mapOf("id" to intArgument(), "optional" to nullableStringArgument(null))
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
         assertWithMessage("Args should not contain optional")
             .that(matchArgs?.containsKey("optional"))
             .isFalse()
@@ -871,10 +830,9 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         assertThat(
-            deepLink.matches(
-                "$DEEP_LINK_EXACT_HTTPS/users?extraParam={extraParam}"
+                deepLink.matches(UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?extraParam={extraParam}"))
             )
-        ).isTrue()
+            .isTrue()
     }
 
     @Test
@@ -884,13 +842,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val myarg = "test"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgumentWithExtraParam.replace("{myarg}", myarg),
-            mapOf("myarg" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgumentWithExtraParam.replace("{myarg}", myarg)),
+                mapOf("myarg" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the argument")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo(myarg)
@@ -903,18 +860,17 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?id={id}&extraParam={extraParam}"
-                .replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    "$DEEP_LINK_EXACT_HTTPS/users?id={id}&extraParam={extraParam}"
+                        .replace("{id}", id.toString())
+                ),
+                mapOf("id" to intArgument())
+            )
 
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -923,13 +879,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?extraParam={extraParam}",
-            mapOf("id" to intArgument(id))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?extraParam={extraParam}"),
+                mapOf("id" to intArgument(id))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the id")
             .that(matchArgs?.containsKey("id"))
             .isFalse()
@@ -940,13 +895,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?id={id}&extraParam={extraParam}",
-            mapOf("myarg" to nullableStringArgument(null))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?id={id}&extraParam={extraParam}"),
+                mapOf("myarg" to nullableStringArgument(null))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the argument")
             .that(matchArgs?.containsKey("myarg"))
             .isFalse()
@@ -958,17 +912,15 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?string={id}"
-                .replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    "$DEEP_LINK_EXACT_HTTPS/users?string={id}".replace("{id}", id.toString())
+                ),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -976,13 +928,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?string={myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("myarg" to nullableStringArgument(null))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("myarg" to nullableStringArgument(null))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the argument")
             .that(matchArgs?.containsKey("myarg"))
             .isFalse()
@@ -994,13 +945,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("id" to intArgument(id))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("id" to intArgument(id))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the id")
             .that(matchArgs?.containsKey("id"))
             .isFalse()
@@ -1012,16 +962,13 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{id}", id.toString())),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -1029,13 +976,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}L"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("myarg" to nullableStringArgument(null))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("myarg" to nullableStringArgument(null))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the argument")
             .that(matchArgs?.containsKey("myarg"))
             .isFalse()
@@ -1047,13 +993,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("id" to intArgument(id))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("id" to intArgument(id))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the id")
             .that(matchArgs?.containsKey("id"))
             .isFalse()
@@ -1066,16 +1011,12 @@ class NavDeepLinkTest {
 
         val first = "Jane"
         val last = "Doe"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{first}", first).replace("{last}", last),
-            mapOf(
-                "first" to stringArgument(),
-                "last" to stringArgument()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{first}", first).replace("{last}", last)),
+                mapOf("first" to stringArgument(), "last" to stringArgument())
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the first name")
             .that(matchArgs?.getString("first"))
             .isEqualTo(first)
@@ -1091,16 +1032,12 @@ class NavDeepLinkTest {
 
         val first = "Jane"
         val last = "Doe"
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf(
-                "first" to stringArgument(first),
-                "last" to stringArgument(last)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("first" to stringArgument(first), "last" to stringArgument(last))
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the first name")
             .that(matchArgs?.containsKey("first"))
             .isFalse()
@@ -1115,13 +1052,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val first = "Jane"
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?name=Jane_",
-            mapOf("first" to stringArgument(), "last" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?name=Jane_"),
+                mapOf("first" to stringArgument(), "last" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the first name")
             .that(matchArgs?.getString("first"))
             .isEqualTo(first)
@@ -1137,13 +1073,12 @@ class NavDeepLinkTest {
 
         val first = ""
         val last = ""
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?name=_",
-            mapOf("first" to stringArgument(), "last" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?name=_"),
+                mapOf("first" to stringArgument(), "last" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the empty first name")
             .that(matchArgs?.getString("first"))
             .isEqualTo(first)
@@ -1157,16 +1092,15 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?name={first}_{last}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf(
-                "first" to nullableStringArgument(null),
-                "last" to nullableStringArgument(null)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf(
+                    "first" to nullableStringArgument(null),
+                    "last" to nullableStringArgument(null)
+                )
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the first name")
             .that(matchArgs?.containsKey("first"))
             .isFalse()
@@ -1181,17 +1115,16 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?productId=wildCardMatch-{id}"
-                .replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    "$DEEP_LINK_EXACT_HTTPS/users?productId=wildCardMatch-{id}"
+                        .replace("{id}", id.toString())
+                ),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -1200,13 +1133,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("id" to intArgument(id))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("id" to intArgument(id))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the id")
             .that(matchArgs?.containsKey("id"))
             .isFalse()
@@ -1217,13 +1149,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?productId=.*-{myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?productId=wildCardMatch-{myarg}",
-            mapOf("myarg" to nullableStringArgument(null))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?productId=wildCardMatch-{myarg}"),
+                mapOf("myarg" to nullableStringArgument(null))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the argument")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo("{myarg}")
@@ -1236,13 +1167,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?productId=.*-",
-            mapOf("id" to intArgument(id))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?productId=.*-"),
+                mapOf("id" to intArgument(id))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the id")
             .that(matchArgs?.containsKey("id"))
             .isFalse()
@@ -1254,17 +1184,15 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?productId=A*B{id}"
-                .replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    "$DEEP_LINK_EXACT_HTTPS/users?productId=A*B{id}".replace("{id}", id.toString())
+                ),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -1273,17 +1201,15 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?productId={id}A*B"
-                .replace("{id}", id.toString()),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    "$DEEP_LINK_EXACT_HTTPS/users?productId={id}A*B".replace("{id}", id.toString())
+                ),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -1292,13 +1218,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val path = "directions"
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?path=go/to/{path}".replace("{path}", path),
-            mapOf("path" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?path=go/to/{path}".replace("{path}", path)),
+                mapOf("path" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the path")
             .that(matchArgs?.getString("path"))
             .isEqualTo(path)
@@ -1310,13 +1235,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val path = "directions"
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("path" to stringArgument(path))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("path" to stringArgument(path))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the path")
             .that(matchArgs?.containsKey("path"))
             .isFalse()
@@ -1327,13 +1251,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?path=go/to/{path}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users",
-            mapOf("path" to nullableStringArgument(null))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users"),
+                mapOf("path" to nullableStringArgument(null))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the path")
             .that(matchArgs?.containsKey("path"))
             .isFalse()
@@ -1345,13 +1268,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?path=go/to/{path}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?path=go/to/",
-            mapOf("path" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?path=go/to/"),
+                mapOf("path" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should not contain the path")
             .that(matchArgs?.getString("path"))
             .isEqualTo("")
@@ -1362,13 +1284,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users#{frag}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users#testFrag",
-            mapOf("frag" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users#testFrag"),
+                mapOf("frag" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the fragment")
             .that(matchArgs?.getString("frag"))
             .isEqualTo("testFrag")
@@ -1379,13 +1300,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?id={id}#{frag}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?id=43#testFrag",
-            mapOf("id" to intArgument(), "frag" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?id=43#testFrag"),
+                mapOf("id" to intArgument(), "frag" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the query")
             .that(matchArgs?.getInt("id"))
             .isEqualTo(43)
@@ -1399,16 +1319,33 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?id={id}#{frag}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users#testFrag",
-            mapOf("id" to nullableStringArgument(), "frag" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users#testFrag"),
+                mapOf("id" to nullableStringArgument(), "frag" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the fragment")
             .that(matchArgs?.getString("frag"))
             .isEqualTo("testFrag")
+    }
+
+    @Test
+    fun deepLinkArgumentMatchEncoded() {
+        val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users/{name}/posts"
+        val deepLink = NavDeepLink(deepLinkArgument)
+
+        val name = "John Doe"
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{name}", UriUtils.encode(name)!!)),
+                mapOf("name" to stringArgument())
+            )
+
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the name")
+            .that(matchArgs?.getString("name"))
+            .isEqualTo(name)
     }
 
     @Test
@@ -1418,21 +1355,17 @@ class NavDeepLinkTest {
 
         val id = 2
         val postId = 42
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument
-                .replace("{id}", id.toString())
-                .replace("{postId}", postId.toString()),
-            mapOf(
-                "id" to intArgument(),
-                "postId" to intArgument()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    deepLinkArgument
+                        .replace("{id}", id.toString())
+                        .replace("{postId}", postId.toString())
+                ),
+                mapOf("id" to intArgument(), "postId" to intArgument())
             )
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
         assertWithMessage("Args should contain the postId")
             .that(matchArgs?.getInt("postId"))
             .isEqualTo(postId)
@@ -1443,7 +1376,7 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users/{id}/posts"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        assertThat(deepLink.matches(deepLinkArgument.replace("{id}", ""))).isTrue()
+        assertThat(deepLink.matches(UriUtils.parse(deepLinkArgument.replace("{id}", "")))).isTrue()
     }
 
     @Test
@@ -1451,8 +1384,7 @@ class NavDeepLinkTest {
         val deepLinkPrefix = "$DEEP_LINK_EXACT_HTTPS/posts/.*"
         val deepLink = NavDeepLink(deepLinkPrefix)
 
-        assertThat(deepLink.matches(deepLinkPrefix.replace(".*", "test")))
-            .isTrue()
+        assertThat(deepLink.matches(UriUtils.parse(deepLinkPrefix.replace(".*", "test")))).isTrue()
     }
 
     @Test
@@ -1460,8 +1392,7 @@ class NavDeepLinkTest {
         val deepLinkWildcard = "$DEEP_LINK_EXACT_HTTPS/posts/.*/new"
         val deepLink = NavDeepLink(deepLinkWildcard)
 
-        assertThat(deepLink.matches(deepLinkWildcard.replace(".*", "test")))
-            .isTrue()
+        assertThat(deepLink.matches(UriUtils.parse(deepLinkWildcard.replace(".*", "test")))).isTrue()
     }
 
     @Test
@@ -1470,15 +1401,14 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkMultiple)
 
         val postId = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkMultiple
-                .replace(".*", "test")
-                .replace("{postId}", postId.toString()),
-            mapOf("postId" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    deepLinkMultiple.replace(".*", "test").replace("{postId}", postId.toString())
+                ),
+                mapOf("postId" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the postId")
             .that(matchArgs?.getInt("postId"))
             .isEqualTo(postId)
@@ -1490,18 +1420,13 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkMultiple)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkMultiple
-                .replace("{id}", id.toString())
-                .replace(".*", "test"),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkMultiple.replace("{id}", id.toString()).replace(".*", "test")),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -1510,16 +1435,13 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            "${DEEP_LINK_EXACT_HTTPS.uppercase()}/users/$id/posts",
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("${DEEP_LINK_EXACT_HTTPS.uppercase()}/users/$id/posts"),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -1528,18 +1450,15 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val id = 2
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument
-                .replace("{id}", id.toString())
-                .replace("users", "Users"),
-            mapOf("id" to intArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args should contain the id")
-            .that(matchArgs?.getInt("id"))
-            .isEqualTo(id)
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(
+                    deepLinkArgument.replace("{id}", id.toString()).replace("users", "Users")
+                ),
+                mapOf("id" to intArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args should contain the id").that(matchArgs?.getInt("id")).isEqualTo(id)
     }
 
     @Test
@@ -1548,20 +1467,16 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkString)
 
         val param = 2
-        val deepLinkUpper = deepLinkString
-            .replace("myParam", "MYPARAM")
-            .replace("{param}", param.toString())
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkUpper,
-            mapOf("param" to intArgument(0))
-        )
+        val deepLinkUpper =
+            deepLinkString.replace("myParam", "MYPARAM").replace("{param}", param.toString())
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkUpper),
+                mapOf("param" to intArgument(0))
+            )
 
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args bundle should be empty")
-            .that(matchArgs?.isEmpty())
-            .isTrue()
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args bundle should be empty").that(matchArgs?.isEmpty()).isTrue()
     }
 
     @Test
@@ -1569,13 +1484,12 @@ class NavDeepLinkTest {
         val deepLinkString = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}"
         val deepLink = NavDeepLink(deepLinkString)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkString,
-            mapOf("myarg" to stringArrayArgument(arrayOf("theArg")))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkString),
+                mapOf("myarg" to stringArrayArgument(arrayOf("theArg")))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         // We allow {argName} values for String types
         assertWithMessage("Args bundle should contain arg value")
             .that(matchArgs?.getStringArray("myarg"))
@@ -1588,13 +1502,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkString)
 
         val value = "%555"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkString.replace("{myarg}", "%25555"),
-            mapOf("myarg" to nullableStringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkString.replace("{myarg}", UriUtils.encode(value)!!)),
+                mapOf("myarg" to nullableStringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the value without additional decoding")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo(value)
@@ -1606,16 +1519,39 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkString)
 
         val value = "some\nthing"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkString.replace("{myarg}", "some%0Athing"),
-            mapOf("myarg" to nullableStringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkString.replace("{myarg}", UriUtils.encode(value)!!)),
+                mapOf("myarg" to nullableStringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the value without additional decoding")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo(value)
+    }
+
+    @Test
+    fun deepLinkMissingRequiredArgumentUnknownDefault() {
+        val deepLink = NavDeepLink("$DEEP_LINK_EXACT_HTTPS?myarg={myarg}")
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(DEEP_LINK_EXACT_HTTPS),
+                // NavArgument with unknown default value
+                mapOf("myarg" to intArgumentUnknownDefault())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+    }
+
+    @Test
+    fun deepLinkMissingNullableArgumentUnknownDefault() {
+        val deepLink = NavDeepLink("$DEEP_LINK_EXACT_HTTPS?myarg={myarg}")
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(DEEP_LINK_EXACT_HTTPS),
+                // NavArgument with unknown default value
+                mapOf("myarg" to nullableStringArgumentUnknownDefault())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
     }
 
     @Test
@@ -1623,17 +1559,13 @@ class NavDeepLinkTest {
         val deepLinkString = "$DEEP_LINK_EXACT_HTTPS/greeting?title={title}&text={text}"
         val deepLink = NavDeepLink(deepLinkString)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/greeting?title=No%20text",
-            mapOf(
-                "title" to stringArgument(),
-                "text" to stringArgument()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/greeting?title=No%20text"),
+                mapOf("title" to stringArgument(), "text" to stringArgument())
             )
-        )
 
-        assertWithMessage("Args should be null")
-            .that(matchArgs)
-            .isNull()
+        assertWithMessage("Args should be null").that(matchArgs).isNull()
     }
 
     @Test
@@ -1641,17 +1573,14 @@ class NavDeepLinkTest {
         val deepLinkString = "$DEEP_LINK_EXACT_HTTPS/greeting?text={text}"
         val deepLink = NavDeepLink(deepLinkString)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/greeting",
-            mapOf("text" to stringArgument("Default greeting"))
-        )
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/greeting"),
+                mapOf("text" to stringArgument("Default greeting"))
+            )
 
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
-        assertWithMessage("Args bundle should be empty")
-            .that(matchArgs?.isEmpty())
-            .isTrue()
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        assertWithMessage("Args bundle should be empty").that(matchArgs?.isEmpty()).isTrue()
     }
 
     @Test
@@ -1660,13 +1589,12 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink(deepLinkArgument)
 
         val args = "test#split"
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{myarg}", args),
-            mapOf("myarg" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{myarg}", args)),
+                mapOf("myarg" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the arg")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo("test")
@@ -1677,13 +1605,12 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?{myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            deepLinkArgument.replace("{myarg}", "name"),
-            mapOf("myarg" to stringArgument())
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse(deepLinkArgument.replace("{myarg}", "name")),
+                mapOf("myarg" to stringArgument())
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         assertWithMessage("Args should contain the arg")
             .that(matchArgs?.getString("myarg"))
             .isEqualTo("name")
@@ -1694,22 +1621,44 @@ class NavDeepLinkTest {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
 
-        val matchArgs = deepLink.getMatchingArguments(
-            "$DEEP_LINK_EXACT_HTTPS/users?myarg=name1&myarg=name2",
-            mapOf("myarg" to stringArrayArgument(null))
-        )
-        assertWithMessage("Args should not be null")
-            .that(matchArgs)
-            .isNotNull()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?myarg=name1&myarg=name2"),
+                mapOf("myarg" to stringArrayArgument(null))
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
         val matchArgsStringArray = matchArgs?.getStringArray("myarg")
-        assertWithMessage("Args list should not be null")
-            .that(matchArgsStringArray)
-            .isNotNull()
+        assertWithMessage("Args list should not be null").that(matchArgsStringArray).isNotNull()
         assertWithMessage("Args should contain first arg")
-            .that(matchArgsStringArray).asList()
+            .that(matchArgsStringArray)
+            .asList()
             .contains("name1")
         assertWithMessage("Args should contain second arg")
-            .that(matchArgsStringArray).asList()
+            .that(matchArgsStringArray)
+            .asList()
+            .contains("name2")
+    }
+
+    @Suppress("UNCHECKED_CAST")
+    @Test
+    fun deepLinkRepeatedQueryParamsMappedToList() {
+        val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}"
+        val deepLink = NavDeepLink(deepLinkArgument)
+
+        val navArg = stringListArgument()
+        val matchArgs =
+            deepLink.getMatchingArguments(
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS/users?myarg=name1&myarg=name2"),
+                mapOf("myarg" to navArg)
+            )
+        assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
+        val matchArgsStringList = navArg.type[matchArgs!!, "myarg"] as List<String>
+        assertWithMessage("Args list should not be null").that(matchArgsStringList).isNotNull()
+        assertWithMessage("Args should contain first arg")
+            .that(matchArgsStringList)
+            .contains("name1")
+        assertWithMessage("Args should contain second arg")
+            .that(matchArgsStringList)
             .contains("name2")
     }
 
@@ -1717,19 +1666,20 @@ class NavDeepLinkTest {
     fun deepLinkNoRepeatedQueryParamsInPattern() {
         val deepLinkArgument = "$DEEP_LINK_EXACT_HTTPS/users?myarg={myarg}&myarg={myarg}"
         val deepLink = NavDeepLink(deepLinkArgument)
-        val message = assertFailsWith<IllegalArgumentException> {
-            // query params are parsed lazily, need to run getMatchingArguments to resolve it
-            deepLink.getMatchingArguments(
-                deepLinkArgument,
-                emptyMap()
+        val message =
+            assertFailsWith<IllegalArgumentException> {
+                    // query params are parsed lazily, need to run getMatchingArguments to resolve
+                    // it
+                    deepLink.getMatchingArguments(UriUtils.parse(deepLinkArgument), emptyMap())
+                }
+                .message
+        assertThat(message)
+            .isEqualTo(
+                "Query parameter myarg must only be present once in $deepLinkArgument. " +
+                    "To support repeated query parameters, use an array type for your " +
+                    "argument and the pattern provided in your URI will be used to " +
+                    "parse each query parameter instance."
             )
-        }.message
-        assertThat(message).isEqualTo(
-            "Query parameter myarg must only be present once in $deepLinkArgument. " +
-                "To support repeated query parameters, use an array type for your " +
-                "argument and the pattern provided in your URI will be used to " +
-                "parse each query parameter instance."
-        )
     }
 
     @Test
@@ -1739,7 +1689,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg="),
                 mapOf("myarg" to stringListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1758,7 +1708,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=&myarg=",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg=&myarg="),
                 mapOf("myarg" to stringListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1777,7 +1727,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                DEEP_LINK_EXACT_HTTPS,
+                UriUtils.parse(DEEP_LINK_EXACT_HTTPS),
                 mapOf("myarg" to stringListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1793,7 +1743,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                DEEP_LINK_EXACT_HTTPS,
+                UriUtils.parse(DEEP_LINK_EXACT_HTTPS),
                 mapOf("myarg" to stringListArgument(listOf("one", "two")))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1806,7 +1756,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=&myarg=two&myarg=three",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg=&myarg=two&myarg=three"),
                 mapOf("myarg" to stringListArgument(listOf("default1", "default2")))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1826,7 +1776,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=one&myarg=&myarg=three",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg=one&myarg=&myarg=three"),
                 mapOf("myarg" to stringListArgument(listOf("default1", "default2")))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1846,7 +1796,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=one&myarg=two&myarg=",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg=one&myarg=two&myarg="),
                 mapOf("myarg" to stringListArgument(listOf("default1", "default2")))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1866,7 +1816,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?"),
                 mapOf("myarg" to stringListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1885,7 +1835,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg="),
                 mapOf("myarg" to intListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1901,7 +1851,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=&myarg=",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg=&myarg="),
                 mapOf("myarg" to intListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1917,7 +1867,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg="),
                 mapOf("myarg" to intListArgument(listOf(1, 2)))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1930,7 +1880,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                DEEP_LINK_EXACT_HTTPS,
+                UriUtils.parse(DEEP_LINK_EXACT_HTTPS),
                 mapOf("myarg" to intListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1946,7 +1896,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                DEEP_LINK_EXACT_HTTPS,
+                UriUtils.parse(DEEP_LINK_EXACT_HTTPS),
                 mapOf("myarg" to intListArgument(listOf(1, 2)))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1959,7 +1909,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=&myarg=2&myarg=3",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg=&myarg=2&myarg=3"),
                 mapOf("myarg" to intListArgument(listOf(-1, -1)))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1979,7 +1929,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=1&myarg=&myarg=3",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg=1&myarg=&myarg=3"),
                 mapOf("myarg" to intListArgument(listOf(-1, -1)))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -1999,7 +1949,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "$DEEP_LINK_EXACT_HTTPS?myarg=1&myarg=2&myarg=",
+                UriUtils.parse("$DEEP_LINK_EXACT_HTTPS?myarg=1&myarg=2&myarg="),
                 mapOf("myarg" to intListArgument(listOf(-1, -1)))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -2019,7 +1969,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "${DEEP_LINK_EXACT_HTTPS}?",
+                UriUtils.parse("${DEEP_LINK_EXACT_HTTPS}?"),
                 mapOf("myarg" to intListArgument())
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -2035,7 +1985,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "${DEEP_LINK_EXACT_HTTPS}?",
+                UriUtils.parse("${DEEP_LINK_EXACT_HTTPS}?"),
                 mapOf("myarg" to intListArgument(listOf(1, 2)))
             )
         assertWithMessage("Args should not be null").that(matchArgs).isNotNull()
@@ -2050,10 +2000,10 @@ class NavDeepLinkTest {
         val deepLink = NavDeepLink.Builder.fromUriPattern<TestClass>(uri).build()
 
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches("http://$uri"))
+            .that(deepLink.matches(UriUtils.parse("http://$uri")))
             .isTrue()
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches("https://$uri"))
+            .that(deepLink.matches(UriUtils.parse("https://$uri")))
             .isTrue()
     }
 
@@ -2066,14 +2016,14 @@ class NavDeepLinkTest {
 
         val expected = "http://$uri/{arg}"
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(expected))
+            .that(deepLink.matches(UriUtils.parse(expected)))
             .isTrue()
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(expected))
+            .that(deepLink.matches(UriUtils.parse(expected)))
             .isTrue()
         val matchArgs =
             deepLink.getMatchingArguments(
-                expected.replace("{arg}", "1"),
+                UriUtils.parse(expected.replace("{arg}", "1")),
                 mapOf("arg" to intArgument())
             )
         assertThat(matchArgs).isNotNull()
@@ -2090,14 +2040,14 @@ class NavDeepLinkTest {
 
         val expected = "http://$uri?arg={arg}"
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(expected))
+            .that(deepLink.matches(UriUtils.parse(expected)))
             .isTrue()
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(expected))
+            .that(deepLink.matches(UriUtils.parse(expected)))
             .isTrue()
         val matchArgs =
             deepLink.getMatchingArguments(
-                expected.replace("{arg}", "2"),
+                UriUtils.parse(expected.replace("{arg}", "2")),
                 mapOf("arg" to intArgumentUnknownDefault())
             )
         assertThat(matchArgs).isNotNull()
@@ -2114,14 +2064,14 @@ class NavDeepLinkTest {
 
         val expected = "http://$uri/{arg2}?arg={arg}"
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(expected))
+            .that(deepLink.matches(UriUtils.parse(expected)))
             .isTrue()
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(expected))
+            .that(deepLink.matches(UriUtils.parse(expected)))
             .isTrue()
         val matchArgs =
             deepLink.getMatchingArguments(
-                expected.replace("{arg}", "1").replace("{arg2}", "false"),
+                UriUtils.parse(expected.replace("{arg}", "1").replace("{arg2}", "false")),
                 mapOf("arg" to intArgumentUnknownDefault(), "arg2" to booleanArgument())
             )
         assertThat(matchArgs).isNotNull()
@@ -2140,7 +2090,7 @@ class NavDeepLinkTest {
         val matchArgs =
             deepLink.getMatchingArguments(
                 // missing required arg in deeplink
-                "http://$uri",
+                UriUtils.parse("http://$uri"),
                 mapOf("arg" to intArgument())
             )
         // should not match since missing default value
@@ -2158,7 +2108,7 @@ class NavDeepLinkTest {
         val matchArgs =
             deepLink.getMatchingArguments(
                 // missing required arg with default value
-                "http://$uri",
+                UriUtils.parse("http://$uri"),
                 mapOf("arg" to intArgumentUnknownDefault())
             )
         // should still match successfully since the string arg has default value
@@ -2175,7 +2125,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "http://$uri",
+                UriUtils.parse("http://$uri"),
                 mapOf("arg" to nullableStringArgument())
             )
         // should not match since missing nullable arg with no default value
@@ -2192,7 +2142,7 @@ class NavDeepLinkTest {
 
         val matchArgs =
             deepLink.getMatchingArguments(
-                "http://$uri",
+                UriUtils.parse("http://$uri"),
                 mapOf("arg" to nullableStringArgumentUnknownDefault())
             )
         assertThat(matchArgs).isNotNull()
@@ -2210,7 +2160,7 @@ class NavDeepLinkTest {
         val route = generateRouteWithArgs(TestClass(null), mapOf(argName to navArg.type))
 
         val matchArgs =
-            deepLink.getMatchingArguments("http://$route", mapOf(argName to navArg))
+            deepLink.getMatchingArguments(UriUtils.parse("http://$route"), mapOf(argName to navArg))
         assertThat(matchArgs).isNotNull()
         assertThat(matchArgs!!.containsKey(argName)).isTrue()
         assertThat(navArg.type[matchArgs, argName]).isNull()
@@ -2228,7 +2178,7 @@ class NavDeepLinkTest {
         val route = generateRouteWithArgs(TestClass(null), mapOf(argName to navArg.type))
 
         val matchArgs =
-            deepLink.getMatchingArguments("http://$route", mapOf(argName to navArg))
+            deepLink.getMatchingArguments(UriUtils.parse("http://$route"), mapOf(argName to navArg))
         assertThat(matchArgs).isNotNull()
         assertThat(matchArgs!!.containsKey(argName)).isTrue()
         assertThat(matchArgs.getStringArray(argName)?.toList()).isEmpty()
@@ -2247,7 +2197,7 @@ class NavDeepLinkTest {
         val route = generateRouteWithArgs(TestClass(null), mapOf(argName to navArg.type))
 
         val matchArgs =
-            deepLink.getMatchingArguments("http://$route", mapOf(argName to navArg))
+            deepLink.getMatchingArguments(UriUtils.parse("http://$route"), mapOf(argName to navArg))
         assertThat(matchArgs).isNotNull()
         assertThat(matchArgs!!.containsKey(argName)).isTrue()
         assertThat(matchArgs.getIntArray(argName)?.toList()).isEmpty()
@@ -2266,7 +2216,7 @@ class NavDeepLinkTest {
         val route = generateRouteWithArgs(TestClass(null), mapOf(argName to navArg.type))
 
         val matchArgs =
-            deepLink.getMatchingArguments("http://$route", mapOf(argName to navArg))
+            deepLink.getMatchingArguments(UriUtils.parse("http://$route"), mapOf(argName to navArg))
         assertThat(matchArgs).isNotNull()
         assertThat(matchArgs!!.containsKey(argName)).isTrue()
         assertThat(matchArgs.getBooleanArray(argName)?.toList()).isEmpty()
@@ -2285,7 +2235,7 @@ class NavDeepLinkTest {
         val route = generateRouteWithArgs(TestClass(null), mapOf(argName to navArg.type))
 
         val matchArgs =
-            deepLink.getMatchingArguments("http://$route", mapOf(argName to navArg))
+            deepLink.getMatchingArguments(UriUtils.parse("http://$route"), mapOf(argName to navArg))
         assertThat(matchArgs).isNotNull()
         assertThat(matchArgs!!.containsKey(argName)).isTrue()
         assertThat(matchArgs.getLongArray(argName)?.toList()).isEmpty()
@@ -2304,7 +2254,7 @@ class NavDeepLinkTest {
         val route = generateRouteWithArgs(TestClass(null), mapOf(argName to navArg.type))
 
         val matchArgs =
-            deepLink.getMatchingArguments("http://$route", mapOf(argName to navArg))
+            deepLink.getMatchingArguments(UriUtils.parse("http://$route"), mapOf(argName to navArg))
         assertThat(matchArgs).isNotNull()
         assertThat(matchArgs!!.containsKey(argName)).isTrue()
         assertThat(matchArgs.getFloatArray(argName)?.toList()).isEmpty()
@@ -2320,11 +2270,11 @@ class NavDeepLinkTest {
 
         val expected = "http://$uri/{arg}/{arg2}"
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(expected))
+            .that(deepLink.matches(UriUtils.parse(expected)))
             .isTrue()
         val matchArgs =
             deepLink.getMatchingArguments(
-                expected.replace("{arg2}", "false"),
+                UriUtils.parse(expected.replace("{arg2}", "false")),
                 mapOf("arg" to intArgument(), "arg2" to booleanArgument())
             )
         assertThat(matchArgs).isNull()
@@ -2340,11 +2290,11 @@ class NavDeepLinkTest {
 
         val expected = "http://$uri/{arg2}?arg={arg}"
         assertWithMessage("No scheme deep links should match http")
-            .that(deepLink.matches(expected))
+            .that(deepLink.matches(UriUtils.parse(expected)))
             .isTrue()
         val matchArgs =
             deepLink.getMatchingArguments(
-                expected.replace("{arg2}", "false"),
+                UriUtils.parse(expected.replace("{arg2}", "false")),
                 mapOf("arg" to nullableStringArgumentUnknownDefault(), "arg2" to booleanArgument())
             )
         assertThat(matchArgs).isNotNull()
