+++ navigation/navigation-runtime/api/desktop/navigation-runtime.api
@@ -18,6 +18,10 @@ public class androidx/navigation/NavController {
 	public fun getNavigatorProvider ()Landroidx/navigation/NavigatorProvider;
 	public fun getPreviousBackStackEntry ()Landroidx/navigation/NavBackStackEntry;
 	public final fun getVisibleEntries ()Lkotlinx/coroutines/flow/StateFlow;
+	public final fun handleDeepLink (Landroidx/navigation/DeepLinkUri;)Z
+	public fun navigate (Landroidx/navigation/NavDeepLinkRequest;)V
+	public fun navigate (Landroidx/navigation/NavDeepLinkRequest;Landroidx/navigation/NavOptions;)V
+	public fun navigate (Landroidx/navigation/NavDeepLinkRequest;Landroidx/navigation/NavOptions;Landroidx/navigation/Navigator$Extras;)V
 	public final fun navigate (Ljava/lang/Object;)V
 	public final fun navigate (Ljava/lang/Object;Landroidx/navigation/NavOptions;)V
 	public final fun navigate (Ljava/lang/Object;Landroidx/navigation/NavOptions;Landroidx/navigation/Navigator$Extras;)V
